Migrate subcommands into the library of the cmsis-cli module
Previously, the subcommands were implemented in the pdsc and cmsis-update modules.
This creates a dependency on clap that cmsis-cffi does not actually use.

This commit moves the dependency on clap into cmsis-cli by moving the implementation
of the subcommands their as well.
