+++ rust/cmsis-cli/Cargo.toml
@@ -18,3 +18,6 @@ path = "../pack-index"
 
 [dependencies.pdsc]
 path = "../pdsc"
+
+[dependencies.utils]
+path = "../utils"
