+++ rust/cmsis-cli/src/main.rs
@@ -7,10 +7,19 @@ extern crate cmsis_update;
 extern crate pack_index;
 extern crate pdsc;
 extern crate failure;
+extern crate cmsis_cli;
 
 use pack_index::config::Config;
-use cmsis_update::{update_args, update_command, install_args, install_command};
-use pdsc::{check_args, check_command, dump_devices_args, dump_devices_command};
+use cmsis_cli::{
+    update_args,
+    update_command,
+    install_args,
+    install_command,
+    check_args,
+    check_command,
+    dump_devices_args,
+    dump_devices_command
+};
 use clap::{Arg, App};
 use slog::Drain;
 use failure::Error;
