+++ rust/cmsis-update/Cargo.toml
@@ -9,7 +9,6 @@ futures-await-async-macro = "0.1.4"
 hyper = "0.11.21"
 hyper-rustls = "0.12.0"
 minidom = "0.5.0"
-clap = "2.19.0"
 slog = "^2"
 slog-term = "^2"
 slog-async = "^2"
