+++ rust/pdsc/Cargo.toml
@@ -6,7 +6,6 @@ authors = ["Jimmy Brisson <theotherjimmy@gmail.com>"]
 [dependencies]
 minidom = "0.5.0"
 slog = "^2"
-clap = "2.19.0"
 serde = "1.0"
 serde_derive = "1.0"
 serde_json = "1.0"
